Splitted tests.
